refactor: 패키지 구조 변경
- model안에 game 패키지 생성: 자동차 경주 관련 클래스 모음
- input, output 패키지 생성: 입력과 출력 분리
