+++ src/main/java/racingcar/model/game/position/Position.java
@@ -1,4 +1,4 @@
-package racingcar.model.position;
+package racingcar.model.game.position;
 
 import java.util.Objects;
