+++ src/main/java/racingcar/model/game/position/Positions.java
@@ -1,4 +1,4 @@
-package racingcar.model.position;
+package racingcar.model.game.position;
 
 import java.util.ArrayList;
 import java.util.List;
