+++ src/main/java/racingcar/view/input/ConsoleInputView.java
@@ -1,6 +1,7 @@
-package racingcar.view;
+package racingcar.view.input;
 
 import camp.nextstep.edu.missionutils.Console;
+import racingcar.view.input.InputView;
 
 public class ConsoleInputView implements InputView {
     @Override
