+++ src/main/java/racingcar/view/input/InputView.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.view.input;
 
 public interface InputView {
     String read();
