+++ src/main/java/racingcar/view/output/ConsoleOutputView.java
@@ -1,8 +1,8 @@
-package racingcar.view;
+package racingcar.view.output;
 
 import java.util.List;
 import java.util.stream.IntStream;
-import racingcar.model.position.Positions;
+import racingcar.model.game.position.Positions;
 import racingcar.support.repeater.StringRepeater;
 
 public class ConsoleOutputView implements OutputView {
