+++ src/test/java/racingcar/model/game/position/HistoryTest.java
@@ -1,4 +1,4 @@
-package racingcar.model.position;
+package racingcar.model.game.position;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -7,6 +7,9 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
+import racingcar.model.game.position.History;
+import racingcar.model.game.position.Position;
+import racingcar.model.game.position.Positions;
 
 @DisplayName("경기 기록 테스트")
 class HistoryTest {
